+++ example/test/unit_test.dart
@@ -1,8 +1,0 @@
-import 'package:test/test.dart';
-
-void main() {
-  test('my first unit test', () {
-    var answer = 11;
-    expect(answer, 11);
-  });
-}