+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.3.+
++ Support db/meter [#41](https://github.com/dooboolab/flutter_sound/pull/41)
+
 ## 1.2.+
 * Fixed sound distorting when playing recorded audio again. Issue [#14](https://github.com/dooboolab/flutter_sound/issues/14).
 * Fixed `seekToPlayer` for android. Issue [#10](https://github.com/dooboolab/flutter_sound/issues/10).
@@ -5,6 +8,9 @@
 + Do not append `tmp` when filePath provided in `ios`.
 + Resolve `regression` issue in `1.2.3` which caused in `1.2.2`.
 + Reduce the size of audio file in `1.2.4`. Related [#26](https://github.com/dooboolab/flutter_sound/issues/26).
++ Fixed `recording` issue in android in `1.2.5`.
++ Changed `seekToPlayer` to place exact `secs` instead adding it.
++ Fix file URI for recording and playing in iOS.
 ## 1.1.+
 * Released 1.1.0 with beautiful logo from mansa.
 * Improved readme.